requirements/requirements.flake.txt
flake8
isort
mypy
pylint
pyright
flake8-builtins
flake8-django
flake8-docstrings
flake8-isort
flake8-mypy
flake8-pylint
flake8-quotes
flake8-todos
django-stubs[compatible-mypy]
djangorestframework-stubs[compatible-mypy]
